cafe-bot/requirements.txt
# Core FastAPI and web server
fastapi==0.115.0
uvicorn[standard]==0.32.0
starlette==0.38.6
pydantic==2.9.2
pydantic_core==2.23.4

# Google Gemini API (LLM)
google-genai==1.56.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2

# Vector search and embeddings (REQUIRED - large packages)
faiss-cpu==1.13.2
sentence-transformers==5.2.0
torch==2.9.1  # Large package (~1.5GB) - required by sentence-transformers
transformers==4.57.3
numpy==2.4.0
scipy==1.16.3
scikit-learn==1.8.0

# HuggingFace and model loading
huggingface-hub==0.36.0
safetensors==0.7.0
tokenizers==0.22.1

# HTTP and networking
httpx==0.28.1
httpcore==1.0.9
httptools==0.7.1
httplib2==0.31.0
requests==2.32.5
websockets==15.0.1
h11==0.16.0
urllib3==2.6.2
certifi==2025.11.12  # SSL certificates
idna==3.11

# Utilities
python-dotenv==1.2.1
PyYAML==6.0.3
tqdm==4.67.1
filelock==3.20.1
fsspec==2025.12.0

# Type hints and validation
typing_extensions==4.15.0
annotated-types==0.7.0

# Async support
anyio==4.12.0
sniffio==1.3.1

# Template engine (FastAPI uses for docs)
Jinja2==3.1.6
MarkupSafe==3.0.3

# Math and scientific computing
mpmath==1.3.0
sympy==1.14.0
joblib==1.5.3
threadpoolctl==3.6.0

# Retry logic
tenacity==9.1.2

# Network utilities
networkx==3.6.1
dnspython==2.8.0  # Used by pymongo if needed

# Build tools (included for compatibility)
setuptools>=65.0.0
packaging==25.0
pyparsing==3.3.1

# Protocol buffers
proto-plus==1.27.0
protobuf==5.29.5

# Authentication
pyasn1==0.6.1
pyasn1_modules==0.4.2
rsa==4.9.1

# URL utilities
uritemplate==4.2.0

# Text processing
regex==2025.11.3
charset-normalizer==3.4.4

# System utilities
cachetools==6.2.4
colorama==0.4.6
distro==1.9.0
click==8.3.1
watchfiles==1.1.1

# MongoDB (optional - only needed for build_cafe_faiss_index.py script)
# Can be removed if you never rebuild the index on Render
pymongo==4.15.5
